provider/perl: skip tests on Windows
It worked for MINGW builds at one point
but it keeps failing now because of perl dependencies
or nvim session issues for tests (named pipes as sockets on Windows?).

+++ test/functional/provider/perl_spec.lua
@@ -20,6 +20,7 @@ before_each(function()
 end)
 
 describe('perl host', function()
+  if helpers.pending_win32(pending) then return end
   teardown(function ()
     os.remove('Xtest-perl-hello.pl')
     os.remove('Xtest-perl-hello-plugin.pl')
